+++ tvx/dummy_testingt.go
@@ -17,3 +17,12 @@ func (d dummyTestingT) Errorf(format string, args ...interface{}) {
 	fmt.Printf(format, args)
 	os.Exit(1)
 }
+
+func (d dummyTestingT) Fatalf(format string, args ...interface{}) {
+	d.Errorf(format, args)
+}
+
+func (d dummyTestingT) Fatal(args ...interface{}) {
+	fmt.Println(args)
+	os.Exit(1)
+}
